feat: improving and better isolating bazel build

+++ demo/03_QuickstartTSP_VNS_BRKGA/TSP-fcore-part4.hpp
@@ -1,6 +1,6 @@
 std::vector<int> frandom(sref<ProblemContext> pTSP) {
   vector<int> v(pTSP->n, -1);  // get information from context
-  for (int i = 0; i < v.size(); i++) v[i] = i;
+  for (int i = 0; i < (int)v.size(); i++) v[i] = i;
   std::random_shuffle(v.begin(), v.end());
   return v;
 }
